WIP - issue because nav & user instantiate eachother
