+++ lib/classes/user_session.js
@@ -1,19 +1,19 @@
+
 import DOMTools from '../tools/dom_tools.js'
 const tool = new DOMTools
 
+import NavBar from './nav_bar.js'
+const nav = new NavBar
+
+
 export default class UserSession {
 
-  // build form
-  // submit form
-  // build page
-  // fill with data
 
   // ---- Logout ----
 
   clearSession(){
     sessionStorage.setItem('api_key', null)
-    // session = false
-    loadNav()
+    nav.loadNav()
   }
 
 
@@ -22,20 +22,20 @@ export default class UserSession {
   registerUser() {
     let content = tool.byId('content')
     tool.clearHTML(content)
-    buildRegisterForm(content)
+    this.buildRegisterForm(content)
   }
 
   buildRegisterForm(element) {
     let form = tool.addSpan(element, 'registerFrom', null)
-    emailField(form)
-    passwordField(form)
-    passwordConfirmationField(form)
-    registerSubmit(form)
+    this.emailField(form)
+    this.passwordField(form)
+    this.passwordConfirmationField(form)
+    this.registerSubmit(form)
   }
 
   makeUser() {
     let user = newUser()
-    makePostRequest(user)
+    this.makePostRequest(user)
   }
 
   newUser() {
@@ -47,26 +47,26 @@ export default class UserSession {
                             'password':               tool.byId('passwordField').value,
                             'password_confirmation':  tool.byId('confirmField').value
                           },
-                  // 'display': 'register'
                 }
     return user
   }
+
+
 
   // ---- Login ----
 
   loginUserForm() {
     let content = tool.byId('content')
     tool.clearHTML(content)
-
     let form = tool.addSpan(content, 'loginFrom', null)
-    emailField(form)
-    passwordField(form)
-    loginSubmit(form)
+    this.emailField(form)
+    this.passwordField(form)
+    this.loginSubmit(form)
   }
 
   loginUser() {
     let user = user()
-    makePostRequest(user)
+    this.makePostRequest(user)
   }
 
   user() {
@@ -82,13 +82,14 @@ export default class UserSession {
     return user
   }
 
+
+
   // ---- API ----
 
   makePostRequest(obj) {
     let verb    = obj['verb']
     let url     = obj['url']
     let body    = obj['body']
-    // let display = obj['display']
 
     var data
     let req = new XMLHttpRequest()
@@ -100,10 +101,9 @@ export default class UserSession {
       let valid = ( stat >= 200 && stat < 300 )
       if( valid ) {
         data = JSON.parse(this.responseText)
-        // assessPost(data, display)
-        newSession(data)
+        this.newSession(data)
       }
-      else { badCredentials() }
+      else { this.badCredentials() }
     }
     var blob = new Blob([ JSON.stringify(body)], {type : 'application/json'});
     req.send( blob )
@@ -113,34 +113,24 @@ export default class UserSession {
     let feedback = tool.byId('feedback')
     tool.clearHTML(feedback)
     let msg = "Sorry, something went wrong."
-    appendText(feedback, msg)
+    tool.appendText(feedback, msg)
   }
 
   newSession(data) {
     let key = data['data']['attributes']['api_key']
     sessionStorage.setItem('api_key', key)
-    // TO DO - Handle this straight from the session storage!
-    // session = true
-    loadNav()
-    nowLoggedIn()
+    nav.loadNav()
+    this.nowLoggedIn()
   }
 
   nowLoggedIn() {
     let feedback = tool.byId('feedback')
     tool.clearHTML(feedback)
     let msg = "You're now logged in, please explore and add favorite locations!"
-    appendText(feedback, msg)
+    tool.appendText(feedback, msg)
     let content = tool.byId('content')
     tool.clearHTML(content)
   }
-
-
-
-
-
-
-
-
 
 
   // ---- HTML Fields ----
@@ -148,7 +138,7 @@ export default class UserSession {
   emailField(form) {
     let email = tool.addDiv(form, 'email', 'row' )
     let title = tool.addSpan(email, null, 'title')
-    appendText(title, "Email")
+    tool.appendText(title, "Email")
     let field = tool.addSpan(email, 'email', 'field')
     field.appendChild( tool.makeInput('text', 'emailField', null) )
   }
@@ -156,7 +146,7 @@ export default class UserSession {
   passwordField(form) {
     let password = tool.addDiv(form, 'password', 'row' )
     let title    = tool.addSpan(password, null, 'title')
-    appendText(title, "Password")
+    tool.appendText(title, "Password")
     let field    = tool.addSpan(password, 'password', 'field')
     field.appendChild( tool.makeInput('password', 'passwordField', null) )
   }
@@ -164,7 +154,7 @@ export default class UserSession {
   passwordConfirmationField(form) {
     let confirm = tool.addDiv(form, 'confrim', 'row' )
     let title   = tool.addSpan(confirm, null, 'title')
-    appendText(title, "Confirm Password")
+    tool.appendText(title, "Confirm Password")
     let field   = tool.addSpan(confirm, 'confirm', 'field')
     field.appendChild( tool.makeInput('password', 'confirmField', null) )
   }
@@ -172,7 +162,7 @@ export default class UserSession {
   registerSubmit(form) {
     let submit = tool.addDiv(form, 'submit', 'row')
     let button = tool.makeInput('submit', 'registerSubmit', null)
-    button.addEventListener('click', function() { makeUser() } )
+    button.addEventListener('click', function() { this.makeUser() } )
     submit.appendChild(button)
     form.appendChild(submit)
   }
@@ -180,10 +170,9 @@ export default class UserSession {
   loginSubmit(form) {
     let submit = tool.addDiv(form, 'submit', 'row')
     let button = tool.makeInput('submit', 'loginSubmit', null)
-    button.addEventListener('click', function() { loginUser() } )
+    button.addEventListener('click', function() { this.loginUser() } )
     submit.appendChild(button)
     form.appendChild(submit)
   }
-
 
 }
