+++ lib/tools/dom_tools.js
@@ -39,4 +39,9 @@ export default class DOMTools {
     return field
   }
 
+  appendText(element, text) {
+    let node = document.createTextNode(text)
+    element.appendChild(node)
+  }
+
 }
